# schema/swagger_old.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Contact Management API",
    "description": "Contact API"
  },
  "basePath": "/api",
  "paths": {
    "/contact": {
      "post": {
        "description": "Add a new Contact",
        "operationId": "addContacts",
        "parameters": [
          {
            "name": "contact",
            "in": "body",
            "description": "The Contact JSON you want to post",
            "schema": {
              "$ref": "#/definitions/Contact"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Contact created successfully",
            "schema": {
              "$ref": "#/definitions/Contact"
            }
          },
          "500": {
            "description": "An unexpected error occured.",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "get": {
        "description": "Returns all contacts from the system",
        "operationId": "getContacts",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "maximum number of results to return",
            "required": false,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "A list of Contacts",
            "schema": {
              "$ref": "#/definitions/Contact"
            }
          }
        }
      }
    },
    "/contacts": {
      "get": {
        "summary": "Gets a Contact",
        "description": "Returns a single contact for its fname, lastname and email.",
        "operationId": "getContact",
        "parameters": [
          {
            "name": "fname",
            "in": "query",
            "required": true,
            "description": "The contact's fname",
            "type": "string"
          },
          {
            "name": "lname",
            "in": "query",
            "required": true,
            "description": "The contact's fname",
            "type": "string"
          },
          {
            "name": "dob",
            "in": "query",
            "required": true,
            "description": "The contact's fname",
            "type": "string",
            "format": "date"
          }
        ],
        "responses": {
          "200": {
            "description": "A Contact",
            "schema": {
              "$ref": "#/definitions/Contact"
            }
          },
          "404": {
            "description": "The Contact does not exists."
          },
          "500": {
            "description": "An unexpected error occured.",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "put": {
        "summary": "Update a Contact in the contact list",
        "description": "Update a Contact in the contact list",
        "operationId": "updateContact",
        "parameters": [
          {
            "name": "fname",
            "in": "query",
            "required": true,
            "description": "The contact's fname",
            "type": "string"
          },
          {
            "name": "lname",
            "in": "query",
            "required": true,
            "description": "The contact's fname",
            "type": "string"
          },
          {
            "name": "dob",
            "in": "query",
            "required": true,
            "description": "The contact's fname",
            "type": "string",
            "format": "date"
          },
          {
            "name": "contact",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Contact"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully updated contact"
          }
        }
      },
      "patch": {
        "summary": "Update partially",
        "description": "support partial updates for contact",
        "operationId": "patchContact",
        "parameters": [
          {
            "name": "fname",
            "in": "query",
            "required": true,
            "description": "The contact's fname",
            "type": "string"
          },
          {
            "name": "lname",
            "in": "query",
            "required": true,
            "description": "The contact's lname",
            "type": "string"
          },
          {
            "name": "dob",
            "in": "query",
            "required": true,
            "description": "The contact's dob",
            "type": "string",
            "format": "date"
          },
          {
            "name": "contact",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/PatchArr"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully updated contact"
          }
        }
      },
      "delete": {
        "summary": "Delete a contact",
        "description": "Delete a contact",
        "operationId": "deleteContact",
        "parameters": [
          {
            "name": "fname",
            "in": "query",
            "required": true,
            "description": "The contact's fname",
            "type": "string"
          },
          {
            "name": "lname",
            "in": "query",
            "required": true,
            "description": "The contact's lname",
            "type": "string"
          },
          {
            "name": "dob",
            "in": "query",
            "required": true,
            "description": "The contact's dob",
            "type": "string",
            "format": "date"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully deleted a person from people list"
          }
        }
      }
    }
  },
  "definitions": {
    "Identification": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "description": "Unique identifier representing a specific Contact."
        },
        "first_name": {
          "type": "string",
          "pattern": "[a-z0-9]{8,64}",
          "minLength": 8,
          "maxLength": 64,
          "description": "First Name."
        },
        "last_name": {
          "type": "string",
          "pattern": "[a-z0-9]{8,64}",
          "minLength": 8,
          "maxLength": 64,
          "description": "Last Name."
        },
        "dob": {
          "type": "string",
          "format": "date",
          "description": "Date of birth"
        },
        "gender": {
          "type": "string",
          "enum": [
            "Male",
            "Female"
          ],
          "description": "Gender."
        },
        "title": {
          "type": "string",
          "enum": [
            "Mr",
            "Mrs",
            "Miss",
            "Dr"
          ],
          "description": "Courtesy Title."
        }
      }
    },
    "Address": {
      "type": "array",
      "items": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Unique identifier representing a specific Address."
          },
          "type": {
            "type": "string",
            "description": "address type e.g home,office"
          },
          "number": {
            "type": "string",
            "description": "house number"
          },
          "street": {
            "type": "string",
            "description": "street address"
          },
          "unit": {
            "type": "string",
            "description": "apartment / unit"
          },
          "city": {
            "type": "string",
            "description": " City"
          },
          "state": {
            "type": "string",
            "description": "State"
          },
          "zipcode": {
            "type": "string",
            "description": "Zip Code"
          }
        }
      }
    },
    "Communication": {
      "type": "array",
      "items": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Unique identifier representing a specific Comunication type."
          },
          "type": {
            "type": "string",
            "description": "Communication methode type e.g email, phone"
          },
          "value": {
            "type": "string",
            "description": "Value of email adress  or phone number"
          },
          "preferred": {
            "type": "string",
            "enum": [
              true,
              false
            ],
            "description": "Preferred Communication method"
          }
        }
      }
    },
    "Contact": {
      "type": "object",
      "properties": {
        "Identification": {
          "$ref": "#/definitions/Identification"
        },
        "Address": {
          "$ref": "#/definitions/Address"
        },
        "Communication": {
          "$ref": "#/definitions/Communication"
        }
      }
    },
    "PatchArr": {
      "type": "array",
      "items": {
        "type": "object",
        "properties": {
          "op": {
            "type": "string",
            "enum": [
              "remove",
              "add",
              "replace"
            ],
            "description": "Allowed operations are add , remove , replace"
          },
          "path": {
            "type": "string",
            "description": "Path of the field to be updated"
          },
          "value": {
            "type": "string",
            "description": "New value"
          }
        }
      }
    },
    "Error": {
      "properties": {
        "code": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    }
  }
}